+++ src/main/java/dev/realme/ash/impl/module/render/ViewClipModule.java
@@ -9,7 +9,7 @@ import dev.realme.ash.impl.event.render.CameraClipEvent;
 
 public class ViewClipModule
 extends ToggleModule {
-    Config<Float> distanceConfig = new NumberConfig<Float>("Distance", "The third-person camera clip distance", Float.valueOf(1.0f), Float.valueOf(3.5f), Float.valueOf(20.0f));
+    Config<Float> distanceConfig = new NumberConfig<Float>("Distance", "The third-person camera clip distance", 1.0f, 3.5f, 20.0f);
 
     public ViewClipModule() {
         super("ViewClip", "Clips your third-person camera through blocks", ModuleCategory.RENDER);
